% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champagne_functions.R
\name{champagne_glass}
\alias{champagne_glass}
\title{Calculate the volume of a champagne glass.}
\usage{
champagne_glass(h, r)
}
\arguments{
\item{h}{Height of the champagne glass.}

\item{r}{Radius of the champagne glass.}
}
\value{
The volume of the champagne glass.
}
\description{
This function calculates the volume of a champagne glass based on its
height and radius.
}
\examples{
champagne_glass(2, 1)
champagne_glass(3, 2)

}
