% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champagne_functions.R
\name{vectorize_function}
\alias{vectorize_function}
\title{Vectorized version of the champagne_glass function using Vectorize.}
\usage{
vectorize_function(h, r)
}
\arguments{
\item{h}{Numeric vector of heights.}

\item{r}{Numeric vector of radii.}
}
\value{
Numeric vector of calculated volumes.
}
\description{
This function calculates the volume of champagne glasses for given heights
and radii using the Vectorize function.
}
\examples{
vectorize_function(c(2, 3, 4), c(1, 2, 3))

}
