% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champagne_functions.R
\name{vectorized_for_loop}
\alias{vectorized_for_loop}
\title{Vectorized version of the champagne_glass function using a for loop.}
\usage{
vectorized_for_loop(heights, radii)
}
\arguments{
\item{heights}{Numeric vector of heights.}

\item{radii}{Numeric vector of radii.}
}
\value{
Numeric vector of calculated volumes.
}
\description{
This function calculates the volume of champagne glasses for given heights
and radii using a vectorized for loop.
}
\examples{
vectorized_for_loop(c(2, 3, 4), c(1, 2, 3))

}
