% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champagne_functions.R
\name{vectorized_purrr_map}
\alias{vectorized_purrr_map}
\title{Vectorized version of the champagne_glass function using purrr map function.}
\usage{
vectorized_purrr_map(heights, radii)
}
\arguments{
\item{heights}{Numeric vector of heights.}

\item{radii}{Numeric vector of radii.}
}
\value{
Numeric vector of calculated volumes.
}
\description{
This function calculates the volume of champagne glasses for given heights
and radii using the purrr map function.
}
\examples{
vectorized_purrr_map(c(2, 3, 4), c(1, 2, 3))

}
