% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champagne_functions.R
\name{vectorized_sapply}
\alias{vectorized_sapply}
\title{Vectorized version of the champagne_glass function using apply functions.}
\usage{
vectorized_sapply(heights, radii)
}
\arguments{
\item{heights}{Numeric vector of heights.}

\item{radii}{Numeric vector of radii.}
}
\value{
Numeric vector of calculated volumes.
}
\description{
This function calculates the volume of champagne glasses for given heights
and radii using apply functions.
}
\examples{
vectorized_sapply(c(2, 3, 4), c(1, 2, 3))

}
